backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
opencv-python==4.5.3.56
numpy==1.21.2
torch
ultralytics==8.0.0
pillow
python-dotenv==0.19.0
openai==0.27.0
google-generativeai==0.3.0 